Completely refactor submitter to allow for generalization.
