+++ indra_reading/scripts/wait_for_complete.py
@@ -70,7 +70,7 @@ if __name__ == '__main__':
     parser = make_parser()
     args = parser.parse_args()
 
-    from indra_reading.scripts.submit_reading_pipeline import wait_for_complete
+    from indra_reading.submission.submitter import wait_for_complete
 
     job_list = None
     if args.job_list is not None:
