+++ indra_reading/util/__init__.py
@@ -7,11 +7,11 @@ def get_s3_job_log_prefix(s3_base, job_name, job_queue="run_db_reading_queue"):
     return s3_root + '%s/' % job_name
 
 
-def get_s3_and_job_prefixes(base_name, group_name=None):
+def get_s3_and_job_prefixes(job_class, base_name, group_name=None):
     """Returns the s3 prefix and job prefix."""
     if not group_name:
         s3_base = base_name
         job_base = base_name
     else:
         s3_base, job_base = [group_name + d + base_name for d in ['/', '_']]
-    return 'reading_results/' + s3_base + '/', job_base
+    return f'{job_class}_results/{s3_base}/{job_base}'
